Update 4.29.16.2
Overhauled entire code logic system by replacing with switch statements. Used the logic of greater than, less than, or equal to, to determine a value for the boolean 'leftStatus.' 0 means the encoders read the same value, 1 means the left motor reads a greater number of counts than the left motor, and 2 means the right motor reads a greater numer of counts than the left motor. UNTESTED AT THE MOMENT - this code is currently untested as of 2:04 PM (this commit), so be advised it may not function as planned